// apps/backoffice/lib/rbac/types.ts
export type Permission =
  // User Management
  | "USER_READ_OWN"
  | "USER_READ_ANY"
  | "USER_UPDATE_OWN"
  | "USER_UPDATE_ANY"
  | "USER_DELETE_OWN"
  | "USER_DELETE_ANY"
  | "USER_CREATE"
  // Content Management
  | "CONTENT_READ_OWN"
  | "CONTENT_READ_ANY"
  | "CONTENT_CREATE"
  | "CONTENT_UPDATE_OWN"
  | "CONTENT_UPDATE_ANY"
  | "CONTENT_DELETE_OWN"
  | "CONTENT_DELETE_ANY"
  | "CONTENT_PUBLISH"
  // File Management
  | "FILE_UPLOAD_OWN"
  | "FILE_UPLOAD_ANY"
  | "FILE_READ_OWN"
  | "FILE_READ_ANY"
  | "FILE_DELETE_OWN"
  | "FILE_DELETE_ANY"
  | "FILE_MANAGE_ORPHANS"
  | "FILE_ADMIN"
  // Settings
  | "SETTINGS_READ"
  | "SETTINGS_UPDATE"
  // Analytics
  | "ANALYTICS_VIEW"
  | "ANALYTICS_EXPORT"
  // Admin
  | "ADMIN_PANEL_ACCESS"
  | "ADMIN_USERS_MANAGE"
  | "ADMIN_ROLES_MANAGE"
  | "ADMIN_PERMISSIONS_MANAGE"
  | "ADMIN_SYSTEM_SETTINGS_MANAGE"
  // Allow dynamic/custom permissions (any string)
  | string